Move watch cache test utils into non test file
